requirements.txt
pandas
numpy
akshare
duckduckgo-search>=5.0
tavily-python
openai
requests
scipy